add readme
TODO:
lstm聚合。
reddit采样sage
