+++ gammagl/models/__init__.py
@@ -2,8 +2,7 @@ from .gcn import GCNModel
 from .gat import GATModel
 from .sgc import SGCModel
 from .gatv2 import GATV2Model
-from .sample_sage import GraphSAGEModel
-from .full_sage import GraphSAGEModel
+from .graphsage import GraphSAGE_Full_Model, GraphSAGE_Sample_Model
 from .gcnii import GCNIIModel
 from .appnp import APPNPModel
 from .gin import GINModel
